requirements.txt
###### Requirements without Version Specifiers ######`
discord.py
lxml
python-dotenv
